Review fixes @claude review
