+++ core/build.gradle
@@ -76,6 +76,10 @@ dependencies {
         exclude group: 'org.apache.logging.log4j', module: 'log4j-core'
     }
     api group: 'io.monix',                          name: 'monix_3',                             version: '3.4.1'
+    // Netty version is security-sensitive here. Bumping it can change FlowControlHandler behavior, which on
+    // ES 9.1+ (auto-read disabled, elastic/elasticsearch#127817) affects the TLS-over-HTTP read flow. 4.1.136
+    // (netty#15053) required the SslHandshakeReadTrigger workaround used by es{version}x SSLNetty4HttpServerTransport
+    // (tech.beshu.ror.utils.SslHandshakeReadTrigger) — re-verify that path before/after changing this version.
     api group: 'io.netty',                          name: 'netty-codec-http2',                   version: '4.1.136.Final'
     api group: 'io.netty',                          name: 'netty-handler',                       version: '4.1.136.Final'
     api group: 'io.netty',                          name: 'netty-transport',                     version: '4.1.136.Final'
